devtools/shared/webconsole/test/chrome/test_basics.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Basic Web Console Actor tests</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="common.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Basic Web Console Actor tests</p>

<script class="testbody" type="text/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();

async function startTest()
{
  removeEventListener("load", startTest);

  let {state, response} = await attachConsoleToTab(["PageError"]);
  is(response.startedListeners.length, 1, "startedListeners.length");
  is(response.startedListeners[0], "PageError", "startedListeners: PageError");

  await closeDebugger(state);
  top.console_ = top.console;
  top.console = { lolz: "foo" };
  ({state, response} = await attachConsoleToTab(["PageError", "ConsoleAPI", "foo"]));

  const startedListeners = response.startedListeners;
  is(startedListeners.length, 2, "startedListeners.length");
  isnot(startedListeners.indexOf("PageError"), -1, "startedListeners: PageError");
  isnot(startedListeners.indexOf("ConsoleAPI"), -1,
        "startedListeners: ConsoleAPI");
  is(startedListeners.indexOf("foo"), -1, "startedListeners: no foo");

  top.console = top.console_;
  response = await state.webConsoleFront.stopListeners(["ConsoleAPI", "foo"]);

  is(response.stoppedListeners.length, 1, "stoppedListeners.length");
  is(response.stoppedListeners[0], "ConsoleAPI", "stoppedListeners: ConsoleAPI");
  await closeDebugger(state);
  ({state, response} = await attachConsoleToTab(["ConsoleAPI"]));

  is(response.startedListeners.length, 1, "startedListeners.length");
  is(response.startedListeners[0], "ConsoleAPI", "startedListeners: ConsoleAPI");

  top.console = top.console_;
  delete top.console_;

  closeDebugger(state, function() {
    SimpleTest.finish();
  });
}

addEventListener("load", startTest);
</script>
</body>
</html>
